Shortest path algorithm is now compatible with our graph data structure
