+++ Assets/MapToolkit-lite.meta~2206de0f46ac46c1bda33261dbbc3a4c13817bf1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3c9e781304b55ee4ba242a143923384f
+timeCreated: 1455919824
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
