--- Assets/Scripts/Map/DijkstraGraph.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4670acc38fdc99c409524079d1483780
-timeCreated: 1455918801
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
